Improved sequence, fixed typo in Edge cleanup portion for cache folder.
Chassis detection and setting of hibernation/fast startup moved to start of script in order to ensure setting prior to other processes. In Prior version, hibernation was disabled at start to clear the outdated hibernation file, then reenabled if device was detected as a laptop/tablet; problem is, if the user closed the script due to disk cleanup time or "patience" then this setting would never be applied.